move directives
