+++ app/javascript/gobierto_dashboards/webapp/components/Widget.vue
@@ -25,7 +25,7 @@
 <script>
 import { GridItem } from "vue-grid-layout";
 import WidgetEditable from "./WidgetEditable";
-import { clickoutside } from "lib/shared"
+import { clickoutside } from "lib/vue/directives"
 
 export default {
   name: "Widget",
