+++ app/javascript/lib/shared/index.js
@@ -30,10 +30,6 @@ import {
   truncate,
   percent
 } from "./modules/vue-filters.js";
-import {
-  VueDirectivesMixin,
-  clickoutside
-} from "./modules/vue-directives.js";
 import { Middleware } from "./modules/middleware.js";
 
 export {
@@ -54,7 +50,5 @@ export {
   money,
   date,
   truncate,
-  percent,
-  VueDirectivesMixin,
-  clickoutside
+  percent
 };
