+++ app/javascript/lib/vue/directives/index.js
@@ -1,14 +1,3 @@
-/**
- * This Vue directives are widely used throughtout the application.
- * They're could be reused importing this module:
- *
- * import { VueDirectivesMixin } from "lib/shared";
- *
- * and then, in a Vue App, like so:
- *
- * mixins: [VueDirectivesMixin],
- *
- */
 export const clickoutside = {
   bind(el, binding, vnode) {
     el.clickOutsideEvent = (event) => {
